refactored and made DateTimeParser
