+++ DoLah/CommandTokenizer.cpp
@@ -47,92 +47,6 @@ std::string CommandTokenizer::findDescription(std::vector<std::string> lineArr) 
     return output;
 }
 
-int CommandTokenizer::getDay(std::string str) {
-    str = std::regex_replace(str, std::regex(dayAppendixPattern), "");
-
-    if (str.length() > 2) {
-        return NULL;
-    } else if (!ParserLibrary::isDecimal(str)) {
-        return NULL;
-    }
-
-    return std::stoi(str);
-}
-
-int CommandTokenizer::getMonth(std::string str) {
-    std::string out;
-    for (size_t m = 0; m < monthPattern.size(); m++) {
-        if (std::regex_match(str, std::regex(monthPattern.at(m), std::regex_constants::icase))) {
-            return m;
-        }
-    }
-
-    return NULL;
-}
-
-int CommandTokenizer::getYear(std::string str) {
-    if (str.length() != 2 && str.length() != 4) {
-        return NULL;
-    } else if (!ParserLibrary::isDecimal(str)) {
-        return NULL;
-    }
-
-    return std::stoi(str);
-}
-
-std::tm CommandTokenizer::toDateFormat(std::vector<std::string> strArr) {
-    time_t t = time(0);
-    struct tm output;
-    localtime_s(&output, &t);
-
-    int day;
-    int month;
-    int year;
-
-    int size = strArr.size();
-    if ((day = getDay(strArr.at(0))) != NULL) {
-        output.tm_mday = day;
-        if (size <= 1) {
-            return output;
-        }
-        if ((month = getMonth(strArr.at(1))) != NULL) {
-            output.tm_mon = month;
-            if (size <= 2) {
-                return output;
-            }
-            if ((year = getYear(strArr.at(2))) != NULL) {
-                output.tm_year = year - 1900;
-                return output;
-            } else {
-                throw std::invalid_argument("");
-            }
-        } else {
-            throw std::invalid_argument("");
-        }
-    } else if ((month = getMonth(strArr.at(0))) != NULL) {
-        output.tm_mon = month;
-        if (size <= 1) {
-            throw std::invalid_argument("");
-        }
-        if ((day = getDay(strArr.at(1))) != NULL) {
-            output.tm_mday = day;
-            if (size <= 2) {
-                return output;
-            }
-            if ((year = getYear(strArr.at(2))) != NULL) {
-                output.tm_year = year - 1900;
-                return output;
-            } else {
-                throw std::invalid_argument("");
-            }
-        } else {
-            throw std::invalid_argument("");
-        }
-    } else {
-        throw std::invalid_argument("");
-    }
-}
-
 std::vector<std::tm> CommandTokenizer::findDate(std::vector<std::string> lineArr) {
     CommandTokenizer ct;
 
@@ -148,7 +62,7 @@ std::vector<std::tm> CommandTokenizer::findDate(std::vector<std::string> lineArr
         if (dateFlag) {
             std::vector<std::string> subVec(lineArr.begin() + i + 1, lineArr.end());
             try {
-                output.push_back(ct.toDateFormat(subVec));
+                output.push_back(DateTimeParser::toDateFormat(subVec));
             } catch (std::invalid_argument e) {
                 throw std::invalid_argument("");
             }
