+++ DoLah/DoLah.vcxproj
@@ -129,11 +129,13 @@
     <ClCompile Include="basic.cpp" />
     <ClCompile Include="CommandParser.cpp" />
     <ClCompile Include="CommandTokenizer.cpp" />
+    <ClCompile Include="DateTimeParser.cpp" />
     <ClCompile Include="ParserLibrary.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="CommandParser.h" />
     <ClInclude Include="CommandTokenizer.h" />
+    <ClInclude Include="DateTimeParser.h" />
     <ClInclude Include="ParserLibrary.h" />
     <ClCompile Include="Models\Calendar.cpp" />
     <ClCompile Include="Models\CalendarHistory.cpp" />
