+++ DoLah/DoLah.vcxproj.filters
@@ -69,6 +69,9 @@
     <ClCompile Include="Models\CalendarHistory.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="DateTimeParser.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="basic.h">
@@ -99,6 +102,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Models\CalendarHistory.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="DateTimeParser.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
