+++ DoLahTest/CommandTokenizerUnitTest.cpp
@@ -8,7 +8,7 @@
 
 #include "CommandTokenizer.h"
 #include "ParserLibrary.h"
-
+#include "DateTimeParser.h"
 
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
@@ -44,57 +44,52 @@ public:
     }
 
     TEST_METHOD(toDateFormatTest1) {
-        CommandTokenizer ct;
         std::string input = "March 17 2015";
         std::string expected = "17/3/2015";
         std::vector<std::string> inputArr = ParserLibrary::explode(input, " ");
-        std::tm time = ct.toDateFormat(inputArr);
+        std::tm time = DateTimeParser::toDateFormat(inputArr);
         std::string actual = std::to_string(time.tm_mday) + "/" + std::to_string(time.tm_mon+1) + "/" + std::to_string(time.tm_year + 1900);
 
         Assert::AreEqual(expected, actual);
     }
 
     TEST_METHOD(toDateFormatTest2) {
-        CommandTokenizer ct;
         std::string input = "March 17";
         std::string expected = "17/3/2015";
         std::vector<std::string> inputArr = ParserLibrary::explode(input, " ");
-        std::tm time = ct.toDateFormat(inputArr);
+        std::tm time = DateTimeParser::toDateFormat(inputArr);
         std::string actual = std::to_string(time.tm_mday) + "/" + std::to_string(time.tm_mon+1) + "/" + std::to_string(time.tm_year + 1900);
 
         Assert::AreEqual(expected, actual);
     }
 
     TEST_METHOD(toDateFormatTest3) {
-        CommandTokenizer ct;
         std::string input = "17th";
         std::string expected = "17/10/2015";
         std::vector<std::string> inputArr = ParserLibrary::explode(input, " ");
-        std::tm time = ct.toDateFormat(inputArr);
+        std::tm time = DateTimeParser::toDateFormat(inputArr);
         std::string actual = std::to_string(time.tm_mday) + "/" + std::to_string(time.tm_mon+1) + "/" + std::to_string(time.tm_year + 1900);
 
         Assert::AreEqual(expected, actual);
     }
 
     TEST_METHOD(toDateFormatTest4) {
-        CommandTokenizer ct;
         std::string input = "March 17th 2015";
         std::string expected = "17/3/2015";
         std::vector<std::string> inputArr = ParserLibrary::explode(input, " ");
-        std::tm time = ct.toDateFormat(inputArr);
+        std::tm time = DateTimeParser::toDateFormat(inputArr);
         std::string actual = std::to_string(time.tm_mday) + "/" + std::to_string(time.tm_mon+1) + "/" + std::to_string(time.tm_year + 1900);
 
         Assert::AreEqual(expected, actual);
     }
 
     TEST_METHOD(toDateFormatTest5) {
-        CommandTokenizer ct;
         std::string input = "March 17th 20150";
         std::vector<std::string> inputArr = ParserLibrary::explode(input, " ");
 
         bool assertion = false;
         try {
-            std::tm time = ct.toDateFormat(inputArr);
+            std::tm time = DateTimeParser::toDateFormat(inputArr);
             std::string actual = std::to_string(time.tm_mday) + "/" + std::to_string(time.tm_mon + 1) + "/" + std::to_string(time.tm_year + 1900);
         } catch (std::invalid_argument e) {
             assertion = true;
@@ -103,13 +98,12 @@ public:
     }
 
     TEST_METHOD(toDateFormatTest6) {
-        CommandTokenizer ct;
         std::string input = "March";
         std::vector<std::string> inputArr = ParserLibrary::explode(input, " ");
 
         bool assertion = false;
         try {
-            std::tm time = ct.toDateFormat(inputArr);
+            std::tm time = DateTimeParser::toDateFormat(inputArr);
             std::string actual = std::to_string(time.tm_mday) + "/" + std::to_string(time.tm_mon + 1) + "/" + std::to_string(time.tm_year + 1900);
         } catch (std::invalid_argument e) {
             assertion = true;
